server/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
websockets>=13.1

# Audio analysis
librosa>=0.10.2
soundfile>=0.12.1
numpy>=1.26.0

# Lyrics
lyricsgenius>=3.0.1

# LLM
google-generativeai>=0.8.0

# Task queue
celery[redis]>=5.4.0
redis>=5.2.0

# Storage
python-dotenv>=1.0.1
aiofiles>=24.1.0

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP client (for external APIs)
httpx>=0.28.0

# Dev / Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
mypy>=1.13.0
